Decomposition of the postgres module
