+++ backend/catalog_service/internal/storage/postgres/dao.go
@@ -13,13 +13,13 @@ type (
 		UpdatedAt    pgtype.Timestamptz `db:"updated_at"`
 	}
 
-	ProductCategory struct {
+	ProductCategoryDAO struct {
 		ProductID  pgtype.UUID        `db:"product_id"`
 		CategoryID pgtype.UUID        `db:"category__id"`
 		AssignedAt pgtype.Timestamptz `db:"assigned_id"`
 	}
 
-	ProductImage struct {
+	ProductImageDAO struct {
 		ImageID   string             `db:"image_id"`
 		ProductID pgtype.UUID        `db:"product__id"`
 		URL       string             `db:"url"`
@@ -28,7 +28,7 @@ type (
 		UpdatedAt pgtype.Timestamptz `db:"updated_at"`
 	}
 
-	Product struct {
+	ProductDAO struct {
 		ProductID   pgtype.UUID        `db:"product_id"`
 		ProductName string             `db:"product_name"`
 		Description string             `db:"description"`
