+++ backend/catalog_service/internal/storage/postgres/postgres.go
@@ -6,8 +6,6 @@ import (
 
 	"github.com/doug-martin/goqu/v9"
 	_ "github.com/doug-martin/goqu/v9/dialect/postgres"
-	"github.com/gofrs/uuid"
-	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
 )
 
@@ -39,30 +37,4 @@ func New(conn string) (*Storage, error) {
 
 func (s *Storage) Stop() {
 	s.DB.Close()
-}
-
-func (s *Storage) WithTx(ctx context.Context, conn *pgxpool.Conn, f func(pgx.Tx) error) error {
-	tx, err := conn.Begin(ctx)
-	if err != nil {
-		return err
-	}
-
-	defer func() {
-		if r := recover(); r != nil {
-			_ = tx.Rollback(ctx)
-		} else if err != nil {
-			_ = tx.Rollback(ctx)
-		}
-	}()
-
-	if err = f(tx); err != nil {
-		return err
-	}
-
-	return tx.Commit(ctx)
-}
-
-func (s *Storage) NewV7() uuid.UUID {
-	uuid, _ := uuid.NewV7()
-	return uuid
 }
